+++ tools/wandbox/tests/test_iuwandbox.py
@@ -26,12 +26,13 @@ except ImportError:
 root = os.path.normpath(os.path.dirname(os.path.abspath(__file__)) + '/../../../')
 fused_src = root + '/fused-src'
 test_src = root + '/test/syntax_tests.cpp'
-test_opt_default = ['--encoding', 'utf-8-sig', '--iutest-use-wandbox-min']
+test_opt_default = ['--encoding', 'utf-8-sig']
 test_opt_nomain = test_opt_default
 test_opt = ['-f"-DIUTEST_USE_MAIN"']
 test_opt.extend(test_opt_default)
 test_opt_verbose = ['--verbose']
 test_opt_dryrun = ['--dryrun']
+test_opt_no_min = ['--iutest-use-wandbox-min']
 
 
 def eprint(*args, **kwargs):
@@ -154,7 +155,7 @@ class iuwandbox_test(iuwandbox_test_base):
             iuwandbox.main()
         output = self.dump()
         self.assertEqual(cm.exception.code, 0, output)
-        self.assertRegex(output, '\[ \s+OK \]')
+        self.assertRegex(output, r'\[ \s+OK \]')
         self.assertFalse('-Wmisleading-indentation' in output)
 
     def test_same_filename(self):
@@ -166,6 +167,18 @@ class iuwandbox_test(iuwandbox_test_base):
         output = self.dump()
         self.assertEqual(cm.exception.code, 0, output)
         self.assertRegex(output, '.*OK.*')
+
+    def test_no_min_run(self):
+        sys.argv[1:] = [test_src]
+        sys.argv.extend(test_opt)
+        sys.argv.extend(test_opt_no_min)
+        print(sys.argv)
+        with self.assertRaises(SystemExit) as cm:
+            iuwandbox.main()
+        output = self.dump()
+        self.assertEqual(cm.exception.code, 0, output)
+        self.assertRegex(output, r'\[ \s+OK \]')
+        self.assertFalse('-Wmisleading-indentation' in output)
 
 
 if __name__ == "__main__":